@@ -1,20 +1,17 @@
 name: fish_redux
 description: Fish Redux is an assembled flutter application framework based on Redux state management.
-version: 0.1.3
+version: 0.3.7
 author: Alibaba Xianyu Team <tino.wjf@alibaba-inc.com>
 homepage: https://github.com/alibaba/fish-redux
 
 environment:
-    sdk: '>=2.0.0-dev.68.0 <3.0.0'
-
+    sdk: ">=2.2.0 <3.0.0"
 dependencies:
     flutter:
         sdk: flutter
-    # test: ^1.5.1
-    test: ^1.3.0
 
 dev_dependencies:
-    # test: ^1.5.1
+    test: ^1.5.1
     flutter_test:
         sdk: flutter
 
