@@ -1,16 +1,5 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide Action, Page;
 
-import 'todo_list_page/page.dart';
+import 'app.dart';
 
-void main() => runApp(MyApp());
-
-class MyApp extends StatefulWidget {
-  @override
-  _MyAppState createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  @override
-  Widget build(BuildContext context) =>
-      MaterialApp(home: ToDoListPage().buildPage(<String, dynamic>{}));
-}
+void main() => runApp(createApp());
