@@ -1,16 +1,13 @@
 import 'package:fish_redux/fish_redux.dart';
 
-enum ToDoEditAction { update, done }
+enum ToDoEditAction { onDone, onChangeTheme }
 
 class ToDoEditActionCreator {
-  static Action update(String name, String desc) {
-    return Action(
-      ToDoEditAction.update,
-      payload: <String, String>{'name': name, 'desc': desc},
-    );
+  static Action onDone() {
+    return const Action(ToDoEditAction.onDone);
   }
 
-  static Action done() {
-    return const Action(ToDoEditAction.done);
+  static Action onChangeTheme() {
+    return const Action(ToDoEditAction.onChangeTheme);
   }
 }
