@@ -2,7 +2,6 @@ import 'package:fish_redux/fish_redux.dart';
 
 import '../todo_list_page/todo_component/component.dart';
 import 'effect.dart';
-import 'reducer.dart';
 import 'state.dart';
 import 'view.dart';
 
@@ -11,10 +10,11 @@ class TodoEditPage extends Page<TodoEditState, ToDoState> {
       : super(
           initState: initState,
           effect: buildEffect(),
-          reducer: buildReducer(),
           view: buildView,
-          middleware: <Middleware<TodoEditState>>[
-            logMiddleware(tag: 'TodoEditPage'),
-          ],
+
+          /// 页面私有AOP，如果需要
+          // middleware: <Middleware<TodoEditState>>[
+          //   logMiddleware(tag: 'TodoEditPage'),
+          // ],
         );
 }
