@@ -1,8 +1,10 @@
 import 'package:fish_redux/fish_redux.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide Action, Page;
+
+import '../global_store/state.dart';
 import '../todo_list_page/todo_component/component.dart';
 
-class TodoEditState implements Cloneable<TodoEditState> {
+class TodoEditState implements GlobalBaseState, Cloneable<TodoEditState> {
   ToDoState toDo;
 
   TextEditingController nameEditController;
@@ -12,13 +14,17 @@ class TodoEditState implements Cloneable<TodoEditState> {
   FocusNode focusNodeDesc;
 
   @override
+  Color themeColor;
+
+  @override
   TodoEditState clone() {
     return TodoEditState()
       ..nameEditController = nameEditController
       ..descEditController = descEditController
-      .. focusNodeName = focusNodeName
-      .. focusNodeDesc = focusNodeDesc
-      ..toDo = toDo;
+      ..focusNodeName = focusNodeName
+      ..focusNodeDesc = focusNodeDesc
+      ..toDo = toDo
+      ..themeColor = themeColor;
   }
 }
 
@@ -29,5 +35,6 @@ TodoEditState initState(ToDoState arg) {
   state.descEditController = TextEditingController(text: arg?.desc);
   state.focusNodeName = FocusNode();
   state.focusNodeDesc = FocusNode();
+
   return state;
 }
