@@ -1,5 +1,5 @@
 import 'package:fish_redux/fish_redux.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide Action, Page;
 
 import 'action.dart';
 import 'state.dart';
@@ -8,6 +8,7 @@ Widget buildView(
     TodoEditState state, Dispatch dispatch, ViewService viewService) {
   return Scaffold(
     appBar: AppBar(
+      backgroundColor: state.themeColor,
       title: const Text('Todo'),
     ),
     body: Container(
@@ -40,6 +41,16 @@ Widget buildView(
               ],
             ),
           ),
+          RaisedButton(
+              padding: const EdgeInsets.only(
+                  left: 20.0, top: 10.0, right: 20.0, bottom: 10.0),
+              color: Colors.blue,
+              child: const Text('Change theme',
+                  style: TextStyle(fontSize: 18),
+                  overflow: TextOverflow.ellipsis),
+              onPressed: () {
+                dispatch(ToDoEditActionCreator.onChangeTheme());
+              }),
           Expanded(
               child: Container(
             margin: const EdgeInsets.only(top: 32.0),
@@ -74,7 +85,7 @@ Widget buildView(
       ),
     ),
     floatingActionButton: FloatingActionButton(
-      onPressed: () => dispatch(ToDoEditActionCreator.done()),
+      onPressed: () => dispatch(ToDoEditActionCreator.onDone()),
       tooltip: 'Done',
       child: const Icon(Icons.done),
     ),
