@@ -9,7 +9,7 @@ import 'dart:async';
 ///    Middleware<T>   ---- AOP
 ///    Store<T>        ---- State management center
 /// 2. Additional abstractions beyond the basic concepts of the ReduxJs community.
-///    Connector<S, P> ---- The connection between big object <T> and small object <P>
+///    Connector<S, P> ---- The connection between big object <S> and small object <P>
 ///    SubReducer<T>   ---- A function that modifies data of partial <T>
 ///    The role of this layer of abstraction
 ///    a. It is obvious that the implementation of combineReducers are decoupled with the grammatical features of JS
@@ -26,16 +26,13 @@ class Action {
   final dynamic payload;
 }
 
-/// Store internal intent type definition.
-enum ActionType { unknown, init, replace, destroy }
-
 /// Definition of the standard Reducer.
 /// If the Reducer needs to respond to the Action, it returns a new state, otherwise it returns the old state.
 typedef Reducer<T> = T Function(T state, Action action);
 
 /// Definition of the standard Dispatch.
 /// Send an "intention".
-typedef Dispatch = void Function(Action action);
+typedef Dispatch = dynamic Function(Action action);
 
 /// Definition of a standard subscription function.
 /// input a subscriber and output an anti-subscription function.
@@ -83,12 +80,22 @@ typedef StoreEnhancer<T> = StoreCreator<T> Function(StoreCreator<T> creator);
 /// Ensure that a T will be cloned at most once during the entire process.
 typedef SubReducer<T> = T Function(T state, Action action, bool isStateCopied);
 
-/// Definition of Connector
-/// 1. How to get an object of type P from an object of type S.
-/// 2. How to synchronize change of an object of type P to an object of type S.
+/// Definition of Connector which connects Reducer<S> with Reducer<P>.
+/// 1. How to get an instance of type P from an instance of type S.
+/// 2. How to synchronize changes of an instance of type P to an instance of type S.
+/// 3. How to clone a new S.
 abstract class AbstractConnector<S, P> {
   P get(S state);
 
-  /// Todo
+  /// For mutable state, there are three abilities needed to be met.
+  ///     1. get: (S) => P
+  ///     2. set: (S, P) => void
+  ///     3. shallow copy: s.clone()
+  ///
+  /// For immutable state, there are two abilities needed to be met.
+  ///     1. get: (S) => P
+  ///     2. set: (S, P) => S
+  ///
+  /// See in [connector].
   SubReducer<S> subReducer(Reducer<P> reducer);
 }
