@@ -12,66 +12,77 @@ void _throwIfNot(bool condition, [String message]) {
   }
 }
 
-Store<T> _createBasicStore<T>(T preloadedState, Reducer<T> reducer) {
-  _throwIfNot(preloadedState != null,
-      'Expected the preloadedState to be non-null value.');
+Store<T> _createStore<T>(final T preloadedState, final Reducer<T> reducer) {
+  _throwIfNot(
+    preloadedState != null,
+    'Expected the preloadedState to be non-null value.',
+  );
 
-  T state = preloadedState;
-  reducer = reducer ?? _noop<T>();
-
-  bool isDispatching = false;
-  final List<_VoidCallback> listeners = <_VoidCallback>[];
-  final StreamController<T> notifyController =
+  final List<_VoidCallback> _listeners = <_VoidCallback>[];
+  final StreamController<T> _notifyController =
       StreamController<T>.broadcast(sync: true);
 
-  final Dispatch dispatch = (Action action) {
-    _throwIfNot(action != null, 'Expected the action to be non-null value.');
-    _throwIfNot(
-        action.type != null, 'Expected the action.type to be non-null value.');
-    _throwIfNot(!isDispatching, 'Reducers may not dispatch actions.');
-
-    try {
-      isDispatching = true;
-      state = reducer(state, action);
-    } finally {
-      isDispatching = false;
-    }
-
-    final List<_VoidCallback> _notifyListeners = listeners.toList(
-      growable: false,
-    );
-    for (_VoidCallback listener in _notifyListeners) {
-      listener();
-    }
-
-    notifyController.add(state);
-  };
+  T _state = preloadedState;
+  Reducer<T> _reducer = reducer ?? _noop<T>();
+  bool _isDispatching = false;
+  bool _isDisposed = false;
 
   return Store<T>()
-    ..getState = (() => state)
-    ..dispatch = dispatch
+    ..getState = (() => _state)
+    ..dispatch = (Action action) {
+      _throwIfNot(action != null, 'Expected the action to be non-null value.');
+      _throwIfNot(action.type != null,
+          'Expected the action.type to be non-null value.');
+      _throwIfNot(!_isDispatching, 'Reducers may not dispatch actions.');
+
+      if (_isDisposed) {
+        return;
+      }
+
+      try {
+        _isDispatching = true;
+        _state = _reducer(_state, action);
+      } finally {
+        _isDispatching = false;
+      }
+
+      final List<_VoidCallback> _notifyListeners = _listeners.toList(
+        growable: false,
+      );
+      for (_VoidCallback listener in _notifyListeners) {
+        listener();
+      }
+
+      _notifyController.add(_state);
+    }
     ..replaceReducer = (Reducer<T> replaceReducer) {
-      reducer = replaceReducer ?? _noop;
-      dispatch(const Action(ActionType.replace));
+      _reducer = replaceReducer ?? _noop;
     }
     ..subscribe = (_VoidCallback listener) {
       _throwIfNot(
-          listener != null, 'Expected the listener to be non-null value.');
-      _throwIfNot(!isDispatching,
-          'You may not call store.subscribe() while the reducer is executing.');
+        listener != null,
+        'Expected the listener to be non-null value.',
+      );
+      _throwIfNot(
+        !_isDispatching,
+        'You may not call store.subscribe() while the reducer is executing.',
+      );
 
-      listeners.add(listener);
+      _listeners.add(listener);
 
       return () {
-        _throwIfNot(!isDispatching,
-            'You may not unsubscribe from a store listener while the reducer is executing.');
-        listeners.remove(listener);
+        _throwIfNot(
+          !_isDispatching,
+          'You may not unsubscribe from a store listener while the reducer is executing.',
+        );
+        _listeners.remove(listener);
       };
     }
-    ..observable = (() => notifyController.stream)
+    ..observable = (() => _notifyController.stream)
     ..teardown = () {
-      listeners.clear();
-      return notifyController.close();
+      _isDisposed = true;
+      _listeners.clear();
+      return _notifyController.close();
     };
 }
 
@@ -79,8 +90,8 @@ Store<T> _createBasicStore<T>(T preloadedState, Reducer<T> reducer) {
 Store<T> createStore<T>(T preloadedState, Reducer<T> reducer,
         [StoreEnhancer<T> enhancer]) =>
     enhancer != null
-        ? enhancer(_createBasicStore)(preloadedState, reducer)
-        : _createBasicStore(preloadedState, reducer);
+        ? enhancer(_createStore)(preloadedState, reducer)
+        : _createStore(preloadedState, reducer);
 
 StoreEnhancer<T> composeStoreEnhancer<T>(List<StoreEnhancer<T>> enhancers) =>
     enhancers == null || enhancers.isEmpty
