@@ -1,24 +1,28 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/widgets.dart' hide Action, Page;
 
 import '../redux/redux.dart';
-import '../redux_component/context.dart';
 import '../redux_component/redux_component.dart';
+import 'adapter.dart';
 
-class RecycleContext<T> extends DefaultContext<T> {
+class RecycleContext<T> extends AdapterContext<T> {
   final Map<Object, List<ContextSys<Object>>> _cachedMap =
       <Object, List<ContextSys<Object>>>{};
   final Map<Object, int> _usedIndexMap = <Object, int>{};
 
   RecycleContext({
-    AbstractLogic<T> factors,
-    PageStore<Object> store,
-    BuildContext buildContext,
-    Get<T> getState,
+    @required AbstractAdapter<T> logic,
+    @required @required Store<Object> store,
+    @required BuildContext buildContext,
+    @required Get<T> getState,
+    @required DispatchBus bus,
+    @required Enhancer<Object> enhancer,
   }) : super(
-          factors: factors,
+          logic: logic,
           store: store,
           buildContext: buildContext,
           getState: getState,
+          bus: bus,
+          enhancer: enhancer,
         );
 
   @override
@@ -67,18 +71,23 @@ class RecycleContext<T> extends DefaultContext<T> {
   }
 }
 
-abstract class RecycleContextMixin<T> implements Logic<T> {
+mixin RecycleContextMixin<T> implements AbstractAdapter<T> {
   @override
-  RecycleContext<T> createContext({
-    PageStore<Object> store,
+  RecycleContext<T> createContext(
+    Store<Object> store,
     BuildContext buildContext,
-    Get<T> getState,
+    Get<T> getState, {
+    @required DispatchBus bus,
+    @required Enhancer<Object> enhancer,
   }) {
+    assert(bus != null && enhancer != null);
     return RecycleContext<T>(
-      factors: this,
+      logic: this,
       store: store,
       buildContext: buildContext,
       getState: getState,
+      bus: bus,
+      enhancer: enhancer,
     );
   }
 }
@@ -89,6 +98,7 @@ ListAdapter combineListAdapters(Iterable<ListAdapter> adapters) {
       .toList(growable: false);
 
   if (list.every((ListAdapter e) => e.itemCount == 1)) {
+    /// The result is AbstractComponent
     return ListAdapter(
       (BuildContext buildContext, final int index) =>
           list[index].itemBuilder(buildContext, 0),
@@ -102,6 +112,7 @@ ListAdapter combineListAdapters(Iterable<ListAdapter> adapters) {
     return count + adapter.itemCount;
   });
 
+  /// The result is AbstractAdapter
   return ListAdapter(
     (BuildContext buildContext, final int index) {
       assert(index >= 0 && index < maxItemCount);
@@ -116,4 +127,30 @@ ListAdapter combineListAdapters(Iterable<ListAdapter> adapters) {
     },
     maxItemCount,
   );
+}
+
+ListAdapter memoizeListAdapter(
+  AbstractAdapterBuilder<Object> result,
+  ContextSys<Object> subCtx,
+) {
+  final Object newState = subCtx.state;
+  if (subCtx.extra['@last-state'] != newState) {
+    subCtx.extra['@last-state'] = newState;
+    subCtx.extra['@last-adapter'] =
+        _memoizeListAdapter(result.buildAdapter(subCtx));
+  }
+
+  return subCtx.extra['@last-adapter'];
+}
+
+ListAdapter _memoizeListAdapter(ListAdapter adapter) {
+  if (adapter.itemCount > 0) {
+    final List<Widget> memoized =
+        List<Widget>.filled(adapter.itemCount, null, growable: false);
+    return ListAdapter((BuildContext context, int index) {
+      return (memoized[index] ??= adapter.itemBuilder(context, index));
+    }, adapter.itemCount);
+  } else {
+    return adapter;
+  }
 }
