@@ -1,5 +1,5 @@
 export 'adapter.dart';
 export 'dynamic_flow_adapter.dart';
 export 'recycle_context.dart';
+export 'source_flow_adapter.dart';
 export 'static_flow_adapter.dart';
-export 'visible_change.dart';
