@@ -1,7 +1,6 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/widgets.dart' hide Action, Page;
 
 import '../redux/redux.dart';
-import '../redux_component/logic.dart';
 import '../redux_component/redux_component.dart';
 import '../utils/utils.dart';
 import 'recycle_context.dart';
@@ -16,10 +15,14 @@ class StaticFlowAdapter<T> extends Logic<T>
     @required List<Dependent<T>> slots,
     Reducer<T> reducer,
     Effect<T> effect,
-    HigherEffect<T> higherEffect,
-    OnError<T> onError,
     ReducerFilter<T> filter,
-    Object Function(T) key,
+
+    /// implement [StateKey] in T instead of using key in Logic.
+    /// class T implements StateKey {
+    ///   Object _key = UniqueKey();
+    ///   Object key() => _key;
+    /// }
+    @deprecated Object Function(T) key,
   })  : assert(slots != null),
         _slots = Collections.compact(slots),
         super(
@@ -32,55 +35,50 @@ class StaticFlowAdapter<T> extends Logic<T>
             )
           ]),
           effect: effect,
-          higherEffect: higherEffect,
-          onError: onError,
           filter: filter,
           dependencies: null,
+          // ignore:deprecated_member_use_from_same_package
           key: key,
         );
 
-  ListAdapter buildAdapter2(PageStore<Object> store, Get<T> getter) {
-    return null;
-  }
-
   @override
-  ListAdapter buildAdapter(
-      T state, Dispatch dispatch, ViewService viewService) {
-    final RecycleContext<T> ctx = viewService;
+  ListAdapter buildAdapter(ContextSys<T> ctx) {
+    final RecycleContext<T> recycleCtx = ctx;
     final List<ListAdapter> adapters = <ListAdapter>[];
 
-    ctx.markAllUnused();
+    recycleCtx.markAllUnused();
     for (int i = 0; i < _slots.length; i++) {
       final Dependent<T> dependent = _slots[i];
-      final Object subObject = dependent.subGetter(ctx.getState)();
+      final Object subObject = dependent.subGetter(recycleCtx.getState)();
       if (!dependent.isComponent()) {
         /// precondition is subObject != null
         if (subObject != null) {
           /// use index of key
-          final ContextSys<Object> subCtx = ctx.reuseOrCreate(i, () {
+          final ContextSys<Object> subCtx = recycleCtx.reuseOrCreate(i, () {
             return dependent.createContext(
-              store: ctx.store,
-              buildContext: ctx.context,
-              getState: ctx.getState,
+              recycleCtx.store,
+              recycleCtx.context,
+              recycleCtx.getState,
+              bus: recycleCtx.bus,
+              enhancer: recycleCtx.enhancer,
             );
           });
-          final ListAdapter subAdapter = dependent.buildAdapter(
-            subCtx.state,
-            subCtx.dispatch,
-            subCtx,
-          );
-          adapters.add(subAdapter);
+
+          /// hack to reduce adapter's rebuilding
+          adapters.add(memoizeListAdapter(dependent, subCtx));
         }
       } else if (subObject != null) {
         adapters.add(ListAdapter((BuildContext buildContext, int index) {
           return dependent.buildComponent(
-            ctx.store,
-            ctx.getState,
+            recycleCtx.store,
+            recycleCtx.getState,
+            bus: recycleCtx.bus,
+            enhancer: recycleCtx.enhancer,
           );
         }, 1));
       }
     }
-    ctx.cleanUnused();
+    recycleCtx.cleanUnused();
 
     return combineListAdapters(adapters);
   }
