@@ -64,7 +64,7 @@ class AutoDispose {
     }
 
     if (newParent != null) {
-      newParent._fields.children ??= Set<AutoDispose>();
+      newParent._fields.children ??= <AutoDispose>{};
       newParent._fields.children.add(this);
     }
     if (oldParent != null) {
@@ -73,19 +73,7 @@ class AutoDispose {
     _fields.parent = newParent;
   }
 
-  AutoDispose registerOnDisposed(void Function() onDisposed) {
-    return AutoDispose()
-      ..setParent(this)
-      ..onDisposed(onDisposed);
-  }
-
-  @deprecated
-  void follow(AutoDispose newParent) {
-    return setParent(newParent);
-  }
-
-  @deprecated
-  AutoDispose follower([void Function() onDisposed]) {
-    return registerOnDisposed(onDisposed);
-  }
+  AutoDispose registerOnDisposed(void Function() onDisposed) => AutoDispose()
+    ..setParent(this)
+    ..onDisposed(onDisposed);
 }
