@@ -1,9 +1,11 @@
 import 'dart:async';
 
-import 'package:flutter/widgets.dart';
+import 'package:flutter/widgets.dart' hide Action, Page;
 
 import '../redux/basic.dart';
+import '../utils/utils.dart';
 import 'basic.dart';
+import 'lifecycle.dart';
 
 AdapterBuilder<T> asAdapter<T>(ViewBuilder<T> view) {
   return (T unstableState, Dispatch dispatch, ViewService service) {
@@ -28,13 +30,19 @@ Effect<T> mergeEffects<T extends K, K>(Effect<K> sup, [Effect<T> sub]) {
   };
 }
 
-//combine & as
-Reducer<T> asReducer<T>(
-        Map<Object, Reducer<T>> map) =>
-    (map == null || map.isEmpty)
-        ? null
-        : (T state, Action action) =>
-            map[action.type]?.call(state, action) ?? state;
+/// combine & as
+/// for action.type which override it's == operator
+Reducer<T> asReducer<T>(Map<Object, Reducer<T>> map) => (map == null ||
+        map.isEmpty)
+    ? null
+    : (T state, Action action) =>
+        map.entries
+            .firstWhere(
+                (MapEntry<Object, Reducer<T>> entry) =>
+                    action.type == entry.key,
+                orElse: () => null)
+            ?.value(state, action) ??
+        state;
 
 Reducer<T> filterReducer<T>(Reducer<T> reducer, ReducerFilter<T> filter) {
   return (reducer == null || filter == null)
@@ -44,20 +52,106 @@ Reducer<T> filterReducer<T>(Reducer<T> reducer, ReducerFilter<T> filter) {
         };
 }
 
+const Object _SUB_EFFECT_RETURN_NULL = Object();
+
 typedef SubEffect<T> = FutureOr<void> Function(Action action, Context<T> ctx);
 
+/// for action.type which override it's == operator
+/// return [UserEffecr]
 Effect<T> combineEffects<T>(Map<Object, SubEffect<T>> map) =>
     (map == null || map.isEmpty)
         ? null
         : (Action action, Context<T> ctx) {
-            final SubEffect<T> subEffect = map[action.type];
-            return subEffect?.call(action, ctx) ?? subEffect != null;
+            final SubEffect<T> subEffect = map.entries
+                .firstWhere(
+                  (MapEntry<Object, SubEffect<T>> entry) =>
+                      action.type == entry.key,
+                  orElse: () => null,
+                )
+                ?.value;
+
+            if (subEffect != null) {
+              return subEffect.call(action, ctx) ?? _SUB_EFFECT_RETURN_NULL;
+            }
+
+            //skip-lifecycle-actions
+            if (action.type is Lifecycle) {
+              return _SUB_EFFECT_RETURN_NULL;
+            }
+
+            /// no subEffect
+            return null;
           };
 
-OnError<T> combineOnErrors<T>(List<OnError<T>> onErrors) =>
-    (Exception exception, Context<T> ctx) =>
-        onErrors.any((OnError<T> onError) => onError(exception, ctx));
+/// return [EffectDispatch]
+Dispatch createEffectDispatch<T>(Effect<T> userEffect, Context<T> ctx) {
+  return (Action action) {
+    final Object result = userEffect?.call(action, ctx);
 
-HigherEffect<T> asHigherEffect<T>(Effect<T> effect) => effect != null
-    ? (Context<T> ctx) => (Action action) => effect(action, ctx)
-    : null;
+    //skip-lifecycle-actions
+    if (action.type is Lifecycle && (result == null || result == false)) {
+      return _SUB_EFFECT_RETURN_NULL;
+    }
+
+    return result;
+  };
+}
+
+/// return [NextDispatch]
+Dispatch createNextDispatch<T>(ContextSys<T> ctx) => (Action action) {
+      ctx.broadcastEffect(action, excluded: true);
+      ctx.store.dispatch(action);
+    };
+
+/// return [Dispatch]
+Dispatch createDispatch<T>(Dispatch onEffect, Dispatch next, Context<T> ctx) =>
+    (Action action) {
+      final Object result = onEffect?.call(action);
+      if (result == null || result == false) {
+        next(action);
+      }
+
+      return result == _SUB_EFFECT_RETURN_NULL ? null : result;
+    };
+
+ViewMiddleware<T> mergeViewMiddleware<T>(List<ViewMiddleware<T>> middleware) {
+  return Collections.reduce<ViewMiddleware<T>>(middleware,
+      (ViewMiddleware<T> first, ViewMiddleware<T> second) {
+    return (AbstractComponent<dynamic> component, Store<T> store) {
+      final Composable<ViewBuilder<dynamic>> inner = first(component, store);
+      final Composable<ViewBuilder<dynamic>> outer = second(component, store);
+      return (ViewBuilder<dynamic> view) {
+        return outer(inner(view));
+      };
+    };
+  });
+}
+
+AdapterMiddleware<T> mergeAdapterMiddleware<T>(
+    List<AdapterMiddleware<T>> middleware) {
+  return Collections.reduce<AdapterMiddleware<T>>(middleware,
+      (AdapterMiddleware<T> first, AdapterMiddleware<T> second) {
+    return (AbstractAdapter<dynamic> component, Store<T> store) {
+      final Composable<AdapterBuilder<dynamic>> inner = first(component, store);
+      final Composable<AdapterBuilder<dynamic>> outer =
+          second(component, store);
+      return (AdapterBuilder<dynamic> view) {
+        return outer(inner(view));
+      };
+    };
+  });
+}
+
+EffectMiddleware<T> mergeEffectMiddleware<T>(
+    List<EffectMiddleware<T>> middleware) {
+  return Collections.reduce<EffectMiddleware<T>>(middleware,
+      (EffectMiddleware<T> first, EffectMiddleware<T> second) {
+    return (AbstractLogic<dynamic> logic, Store<T> store) {
+      final Composable<Effect<dynamic>> inner = first(logic, store);
+      final Composable<Effect<dynamic>> outer = second(logic, store);
+      return (Effect<dynamic> effect) {
+        return outer(inner(effect));
+      };
+    };
+  });
+}
