@@ -1,25 +1,46 @@
+import 'package:flutter/scheduler.dart';
+
 import '../redux/redux.dart';
 
 enum Lifecycle {
+  /// componenmt(page) or adapter receives the following events
   initState,
   didChangeDependencies,
   build,
-
+  reassemble,
   didUpdateWidget,
   deactivate,
   dispose,
+  // didDisposed,
 
-  //adapter
+  /// Only a adapter mixin VisibleChangeMixin will receive appear & disappear events.
+  /// class MyAdapter extends Adapter<T> with VisibleChangeMixin<T> {
+  ///   MyAdapter():super(
+  ///     ///
+  ///   );
+  /// }
   appear,
   disappear,
+
+  /// Only a componenmt(page) or adapter mixin WidgetsBindingObserverMixin will receive didChangeAppLifecycleState event.
+  /// class MyComponent extends Component<T> with WidgetsBindingObserverMixin<T> {
+  ///   MyComponent():super(
+  ///     ///
+  ///   );
+  /// }
+  didChangeAppLifecycleState,
 }
 
 class LifecycleCreator {
   static Action initState() => const Action(Lifecycle.initState);
 
-  static Action build() => const Action(Lifecycle.build);
+  static Action build(String name) => Action(Lifecycle.build, payload: name);
+
+  static Action reassemble() => const Action(Lifecycle.reassemble);
 
   static Action dispose() => const Action(Lifecycle.dispose);
+
+  // static Action didDisposed() => const Action(Lifecycle.didDisposed);
 
   static Action didUpdateWidget() => const Action(Lifecycle.didUpdateWidget);
 
@@ -32,4 +53,7 @@ class LifecycleCreator {
 
   static Action disappear(int index) =>
       Action(Lifecycle.disappear, payload: index);
+
+  static Action didChangeAppLifecycleState(AppLifecycleState state) =>
+      Action(Lifecycle.didChangeAppLifecycleState, payload: state);
 }
