@@ -1,24 +1,6 @@
-import 'mutable.dart';
+import '../redux/redux.dart';
+
 import 'op_mixin.dart';
-
-/// use ConnOp<T, P> instead of Connector<T, P>
-@deprecated
-class Connector<T, P> extends MutableConn<T, P> {
-  final P Function(T) _getter;
-  final void Function(T, P) _setter;
-
-  const Connector({
-    P Function(T) get,
-    void Function(T, P) set,
-  })  : _getter = get,
-        _setter = set;
-
-  @override
-  P get(T state) => _getter(state);
-
-  @override
-  void set(T state, P subState) => _setter(state, subState);
-}
 
 class ConnOp<T, P> extends MutableConn<T, P> with ConnOpMixin<T, P> {
   final P Function(T) _getter;
