@@ -1,8 +1,7 @@
-import '../redux/basic.dart';
-import '../redux_component/basic.dart';
-import '../redux_component/dependent.dart';
-import '../redux_component/logic.dart';
+import '../redux/redux.dart';
+import '../redux_component/redux_component.dart';
 
 mixin ConnOpMixin<T, P> on AbstractConnector<T, P> {
-  Dependent<T> operator +(Logic<P> logic) => createDependent<T, P>(this, logic);
+  Dependent<T> operator +(AbstractLogic<P> logic) =>
+      createDependent<T, P>(this, logic);
 }
