@@ -1,6 +1,9 @@
-import 'mutable.dart';
+import 'package:flutter/foundation.dart';
 
-bool _listEquals<E>(List<E> list1, List<E> list2) {
+import '../redux/redux.dart';
+import 'op_mixin.dart';
+
+bool _listEquals(List<dynamic> list1, List<dynamic> list2) {
   if (identical(list1, list2)) {
     return true;
   }
@@ -12,16 +15,24 @@ bool _listEquals<E>(List<E> list1, List<E> list2) {
     return false;
   }
   for (int i = 0; i < length; i++) {
-    if (list1[i] != list2[i]) {
+    final dynamic e1 = list1[i], e2 = list2[i];
+    if (e1 != e2) {
+      if (e1 is List && e1.runtimeType == e2?.runtimeType) {
+        if (!_listEquals(e1, e2)) {
+          return false;
+        }
+      }
       return false;
     }
   }
   return true;
 }
 
-abstract class _BasicReselect<T, P> extends MutableConn<T, P> {
+abstract class _BasicReselect<T, P> extends MutableConn<T, P>
+    with ConnOpMixin<T, P> {
   List<dynamic> _subsCache;
   P _pCache;
+  bool _hasBeenCalled = false;
 
   List<dynamic> getSubs(T state);
 
@@ -30,9 +41,10 @@ abstract class _BasicReselect<T, P> extends MutableConn<T, P> {
   @override
   P get(T state) {
     final List<dynamic> subs = getSubs(state);
-    if (!_listEquals<dynamic>(subs, _subsCache)) {
+    if (!_hasBeenCalled || !_listEquals(subs, _subsCache)) {
       _subsCache = subs;
-      return _pCache = reduceSubs(_subsCache);
+      _pCache = reduceSubs(_subsCache);
+      _hasBeenCalled = true;
     }
     return _pCache;
   }
@@ -40,32 +52,32 @@ abstract class _BasicReselect<T, P> extends MutableConn<T, P> {
 
 abstract class Reselect1<T, P, K0> extends _BasicReselect<T, P> {
   K0 getSub0(T state);
-  P cumputed(K0 state);
+  P computed(K0 state);
 
   @override
   List<dynamic> getSubs(T state) => <dynamic>[getSub0(state)];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(cumputed, list);
+  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
 }
 
 abstract class Reselect2<T, P, K0, K1> extends _BasicReselect<T, P> {
   K0 getSub0(T state);
   K1 getSub1(T state);
-  P cumputed(K0 sub0, K1 sub1);
+  P computed(K0 sub0, K1 sub1);
 
   @override
   List<dynamic> getSubs(T state) => <dynamic>[getSub0(state), getSub1(state)];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(cumputed, list);
+  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
 }
 
 abstract class Reselect3<T, P, K0, K1, K2> extends _BasicReselect<T, P> {
   K0 getSub0(T state);
   K1 getSub1(T state);
   K2 getSub2(T state);
-  P cumputed(K0 sub0, K1 sub1, K2 sub2);
+  P computed(K0 sub0, K1 sub1, K2 sub2);
 
   @override
   List<dynamic> getSubs(T state) => <dynamic>[
@@ -75,7 +87,7 @@ abstract class Reselect3<T, P, K0, K1, K2> extends _BasicReselect<T, P> {
       ];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(cumputed, list);
+  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
 }
 
 abstract class Reselect4<T, P, K0, K1, K2, K3> extends _BasicReselect<T, P> {
@@ -83,7 +95,7 @@ abstract class Reselect4<T, P, K0, K1, K2, K3> extends _BasicReselect<T, P> {
   K1 getSub1(T state);
   K2 getSub2(T state);
   K3 getSub3(T state);
-  P cumputed(K0 sub0, K1 sub1, K2 sub2, K3 sub3);
+  P computed(K0 sub0, K1 sub1, K2 sub2, K3 sub3);
 
   @override
   List<dynamic> getSubs(T state) => <dynamic>[
@@ -94,7 +106,7 @@ abstract class Reselect4<T, P, K0, K1, K2, K3> extends _BasicReselect<T, P> {
       ];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(cumputed, list);
+  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
 }
 
 abstract class Reselect5<T, P, K0, K1, K2, K3, K4>
@@ -104,7 +116,7 @@ abstract class Reselect5<T, P, K0, K1, K2, K3, K4>
   K2 getSub2(T state);
   K3 getSub3(T state);
   K4 getSub4(T state);
-  P cumputed(K0 sub0, K1 sub1, K2 sub2, K3 sub3, K4 sub4);
+  P computed(K0 sub0, K1 sub1, K2 sub2, K3 sub3, K4 sub4);
 
   @override
   List<dynamic> getSubs(T state) => <dynamic>[
@@ -116,7 +128,7 @@ abstract class Reselect5<T, P, K0, K1, K2, K3, K4>
       ];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(cumputed, list);
+  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
 }
 
 abstract class Reselect6<T, P, K0, K1, K2, K3, K4, K5>
@@ -127,7 +139,7 @@ abstract class Reselect6<T, P, K0, K1, K2, K3, K4, K5>
   K3 getSub3(T state);
   K4 getSub4(T state);
   K5 getSub5(T state);
-  P cumputed(K0 sub0, K1 sub1, K2 sub2, K3 sub3, K4 sub4, K5 sub5);
+  P computed(K0 sub0, K1 sub1, K2 sub2, K3 sub3, K4 sub4, K5 sub5);
 
   @override
   List<dynamic> getSubs(T state) => <dynamic>[
@@ -140,12 +152,35 @@ abstract class Reselect6<T, P, K0, K1, K2, K3, K4, K5>
       ];
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(cumputed, list);
+  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
 }
 
 abstract class Reselect<T, P> extends _BasicReselect<T, P> {
-  P cumputed(List<dynamic> list);
+  P computed(List<dynamic> list);
 
   @override
-  P reduceSubs(List<dynamic> list) => Function.apply(cumputed, list);
+  P reduceSubs(List<dynamic> list) => Function.apply(computed, list);
+}
+
+/// issue [https://github.com/alibaba/fish-redux/issues/482]
+mixin ReselectMixin<T, P> on MutableConn<T, P> {
+  List<dynamic> _cachedFactors;
+  P _cachedResult;
+  bool _hasBeenCalled = false;
+
+  P computed(T state);
+
+  List<dynamic> factors(T state) => <dynamic>[state];
+
+  @mustCallSuper
+  @override
+  P get(T state) {
+    final List<dynamic> newFactors = factors(state);
+    if (!_hasBeenCalled || !_listEquals(newFactors, _cachedFactors)) {
+      _cachedFactors = newFactors.toList(growable: false);
+      _cachedResult = computed(state);
+      _hasBeenCalled = true;
+    }
+    return _cachedResult;
+  }
 }
