@@ -1,5 +1,5 @@
 import 'package:fish_redux/fish_redux.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide Action, Page;
 
 import '../test_base.dart';
 import 'action.dart';
@@ -171,9 +171,6 @@ dynamic toDoListEffectAsync(Action action, Context<ToDoList> ctx) {
 
   return null;
 }
-
-OnAction toDoListHigherEffect(Context<ToDoList> ctx) =>
-    (Action action) => toDoListEffect(action, ctx);
 
 ToDoList toDoListReducer(ToDoList state, Action action) {
   print('onReduce:${action.type}');
