@@ -24,7 +24,7 @@ dynamic toDoListEffectAsync(Action action, Context<ToDoList> ctx) {
   return null;
 }
 
-OnAction toDoListHigherEffect(Context<ToDoList> ctx) =>
+Dispatch toDoListHigherEffect(Context<ToDoList> ctx) =>
     (Action action) => toDoListEffect(action, ctx);
 
 ToDoList toDoListReducer(ToDoList state, Action action) {
@@ -51,16 +51,20 @@ class ToDoComponent2 extends ToDoComponent {}
 class ToDoComponent3 extends ToDoComponent {}
 
 final testAdapter = TestStaticFlowAdapter<ToDoList>(slots: [
-  ToDoComponent0().asDependent(Connector<ToDoList, Todo>(
-      get: (toDoList) => toDoList.list[0],
-      set: (toDoList, toDo) => toDoList.list[0] = toDo)),
-  ToDoComponent1().asDependent(Connector<ToDoList, Todo>(
-      get: (toDoList) => toDoList.list[1],
-      set: (toDoList, toDo) => toDoList.list[1] = toDo)),
-  ToDoComponent2().asDependent(Connector<ToDoList, Todo>(
-      get: (toDoList) => toDoList.list[2],
-      set: (toDoList, toDo) => toDoList.list[2] = toDo)),
-  ToDoComponent3().asDependent(Connector<ToDoList, Todo>(
-      get: (toDoList) => toDoList.list[3],
-      set: (toDoList, toDo) => toDoList.list[3] = toDo))
+  ConnOp<ToDoList, Todo>(
+          get: (toDoList) => toDoList.list[0],
+          set: (toDoList, toDo) => toDoList.list[0] = toDo) +
+      ToDoComponent0(),
+  ConnOp<ToDoList, Todo>(
+          get: (toDoList) => toDoList.list[1],
+          set: (toDoList, toDo) => toDoList.list[1] = toDo) +
+      ToDoComponent1(),
+  ConnOp<ToDoList, Todo>(
+          get: (toDoList) => toDoList.list[2],
+          set: (toDoList, toDo) => toDoList.list[2] = toDo) +
+      ToDoComponent2(),
+  ConnOp<ToDoList, Todo>(
+          get: (toDoList) => toDoList.list[3],
+          set: (toDoList, toDo) => toDoList.list[3] = toDo) +
+      ToDoComponent3()
 ], reducer: toDoListReducer, effect: toDoListEffect);
